remove controllers with machines
